refactor: keep SavedEntry local to each app, not in @aggy/lib
SavedEntry is a storage/persistence concept, not domain logic.
Web defines it in opfs.ts; desktop defines it in storage.ts.
